services/meal-planner-service/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
python-dotenv==1.0.0
httpx==0.24.1
pydantic==2.0.2
google-generativeai==0.3.1
langchain==0.1.0
langchain-google-genai==0.0.5
numpy==1.24.3
scikit-learn==1.3.0
pgvector==0.2.3
langchain_community==0.0.13
python-jose==3.3.0